+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Fixed T-Deck trackball spamming inputs
 - Removed limit of 50 for 'scanap' to prevent getting rid of early entries
 - Changed "Unknown command" to "Unsupported command" in CLI error messages for better UX.
+- Improved Cardputer charging detection
 
 ## Revival v1.9